% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{battleResults}
\alias{battleResults}
\title{Results from a battle}
\description{
a list with members
\describe{
 \item{unitsAttacker}{character() vector with the remaining units for attacker.}
 \item{unitsDefender}{character() vector with the remaining units for defender.}
 \item{rounds}{integer() the number of rounds in the battler}
 \item{attackerCost}{numeric() The cost of units lost for attacker}
 \item{defenderCost}{numeric() The cost of units lost for defender}
}
}
