% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{calculatePosteriorDistribution}
\alias{calculatePosteriorDistribution}
\title{posterior distribution}
\usage{
calculatePosteriorDistribution(simulationResults, side = "attacker")
}
\arguments{
\item{simulationResults, }{formatted as: \code{\link[aaSimulator]{simulationResults}}}

\item{side}{character(), side to calculate distribution for, either 'attacker' or 'defender'}
}
\value{
data.table with posterior probabilities of survival for each unit in ool (in original order). Each replicate is represented by one row.
}
\description{
Calculate posterior distribution for battle
}
