% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{calculateStats}
\alias{calculateStats}
\title{Calculate battle statistics}
\usage{
calculateStats(simulationResults)
}
\arguments{
\item{simulationResults, }{formatted as: \code{\link[aaSimulator]{simulationResults}}}
}
\value{
\code{\link[aaSimulator]{simulationStats}}
}
\description{
Calculates battle statistics for simulation
}
