% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{collapseOol}
\alias{collapseOol}
\title{Collapse OOL}
\usage{
collapseOol(ool)
}
\arguments{
\item{ool}{order of loss specification formatted as \code{\link[aaSimulator]{ool}}}
}
\value{
order of loss specification formatted as \code{\link[aaSimulator]{prefixedOol}}
}
\description{
Collapse order of loss specification to prefixed form.
}
