% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{expandPrefixedOOL}
\alias{expandPrefixedOOL}
\title{Expand prefixed OOL}
\usage{
expandPrefixedOOL(prefixedOol)
}
\arguments{
\item{prefixedOol}{order of loss specification formatted as \code{\link[aaSimulator]{prefixedOol}}}
}
\value{
order of loss specification formatted as \code{\link[aaSimulator]{ool}}
}
\description{
Expand order of loss specification o. prefixed form.
}
