% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_dice.R
\name{inspect_dice}
\alias{inspect_dice}
\title{Inspect dice}
\usage{
inspect_dice(iterations, ones = 0, twos = 0, threes = 0, fours = 0, fives = 0)
}
\arguments{
\item{iterations}{integer() The number of iterations to run.}

\item{ones}{parameter passed to \code{\link[aaSimulator]{roll}}.}

\item{twos}{parameter passed to \code{\link[aaSimulator]{roll}}.}

\item{threes}{parameter passed to \code{\link[aaSimulator]{roll}}.}

\item{fours}{parameter passed to \code{\link[aaSimulator]{roll}}.}

\item{fives}{parameter passed to \code{\link[aaSimulator]{roll}}.}
}
\description{
Inspects dice for fairness and plots distribution of results.
}
\details{
Rolls the given dice configuration repeatedly using \code{\link[aaSimulator]{roll}}.
}
