% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhtr_units_doc.R
\docType{data}
\name{lhtr2_units}
\alias{lhtr2_units}
\title{Table of units allowed in LHTR 2.0}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 21 rows and 8 columns.
}
\usage{
data(lhtr2_units)
}
\description{
Formatted as \code{\link[aaSimulator]{unitTable}}
}
\keyword{datasets}
