% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{makeBattleStatPlot}
\alias{makeBattleStatPlot}
\title{Make stat table plot}
\usage{
makeBattleStatPlot(simulationResults)
}
\arguments{
\item{simulationResults, }{formatted as: \code{\link[aaSimulator]{simulationResults}}}
}
\value{
A gtable, see \code{\link[gridExtra]{tableGrob}}.
}
\description{
Makes a table over some key stats summarising a battle simulation
}
\examples{
 sim <- simulateBattles("30 inf 11 arm 5 ftr arm",
        "40 inf 3 ftr",
        iterations = 1000,
        replications = 2)
 makeBattleStatPlot(sim)
}
