% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{makeCostPlot}
\alias{makeCostPlot}
\title{Make plot of costs}
\usage{
makeCostPlot(
  simulationResults,
  attackCol = "red",
  defenceCol = "blue",
  legend = T
)
}
\arguments{
\item{simulationResults, }{formatted as: \code{\link[aaSimulator]{simulationResults}}}

\item{attackCol}{color to use when side is 'attacker'}

\item{defenceCol}{color to use when side is 'defender'}

\item{legend}{logical() whether to add a legend to the plot}
}
\value{
ggplot object
}
\description{
Plots the cumulative probabilities of costs for a battle simulation
}
\examples{
 sim <- simulateBattles("30 inf 11 arm 5 ftr arm",
        "40 inf 3 ftr",
        iterations = 1000,
        replications = 2)
 makeCostPlot(sim)
}
