% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{makePosteriorDistributionPlot}
\alias{makePosteriorDistributionPlot}
\title{plot posterior distribution}
\usage{
makePosteriorDistributionPlot(
  simulationResults,
  side = "attacker",
  attackCol = "red",
  defenceCol = "blue"
)
}
\arguments{
\item{simulationResults, }{formatted as: \code{\link[aaSimulator]{simulationResults}}}

\item{side}{character(), side to calculate distribution for, either 'attacker' or 'defender'}

\item{attackCol}{color to use when side is 'attacker'}

\item{defenceCol}{color to use when side is 'defender'}
}
\value{
ggplot object
}
\description{
Calculate and plot posterior distribution for battle
}
\details{
Plotter error bars are the ranges of the reported probabilities across replicates.
}
