% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{ool}
\alias{ool}
\title{OOL order of loss specification
A character vector specifying the units participating in a battle, ordered by their priority (first units lost first).
May also contain virtual units, that may be special abilities or control directives to the battle simulator.}
\description{
Exact notation is determined by the battle function used. But for purposes of harmonization, the following notation is encouraged from common units:
}
\details{
\describe{
  \item{inf}{Infantry}
  \item{art}{Artillery}
  \item{arm}{Tank}
  \item{ftr}{Fighter}
  \item{bmb}{Bomber}
  \item{trn}{Transport}
  \item{sub}{Submarine}
  \item{dd}{Destroyer}
  \item{ac}{Aircraft Carrier}
  \item{bb}{Battleship}
}

 Some common virtual units:
\describe{
  \item{AA}{Anti-aircraft Gun. Considered a virtual unit for rule sets where it can not be taken as casualty}
  \item{BBomb}{Offshore bombardment from Battleship, supporting amphibious assults.}
  \item{BBx}{The first hit of a battleship for rulesets with two-hit capabilities. Modelled as a virtual unit with 0 in attack and defense.}
  \item{SUBM}{Flags that all attacking submarines should submerge when possible, after all units preceeding this virtual unit are lost.}
  \item{RET}{Flags that attacker should retreat at first opportunity when all units preceeding this virtual unit are lost.}
}
}
