% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{optimizeUnits}
\alias{optimizeUnits}
\title{Optimize units}
\usage{
optimizeUnits(
  cost,
  iterations = 1000,
  replications = 3,
  attacker = NULL,
  defender = NULL,
  units = c(),
  unittable = lhtr2_units
)
}
\arguments{
\item{cost}{The cost to optimize for, unit configurations must not exceed this cost}

\item{iterations}{the number of iterations to run for each unit configuration}

\item{replications}{the number of replicates to run for each unit configuration}

\item{attacker}{ool for attacker, NULL if attacking units are to be optimized}

\item{defender}{ool for defender, NULL if defending units are to be optimized}

\item{units}{the units in order of loss that should be sampled for optimization, formatted as \code{\link[aaSimulator]{ool}}.}

\item{unittable}{table of unit properties, formatted as \code{\link[aaSimulator]{unitTable}}}
}
\value{
list containing the optimal unit configurations. That is the unit configuration that are optimal for the given cost on average,
 and all unit configurations where some replicate peforms at least as good as the optimum average.
}
\description{
Optimize unit configuration for fixed costs.
}
\details{
'units' denote order of unit groups, so that units=c("inf", "arm"), will try all combinations
 of "inf" and "arm" allowed by the parameter 'cost', but always with all "inf" preceeding all "arm".
}
