% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1942.R
\name{play_1942se_battle}
\alias{play_1942se_battle}
\title{Play one battle}
\usage{
play_1942se_battle(
  oolAttacker,
  oolDefender,
  retreat = NULL,
  verbose = F,
  suppressChecks = F
)
}
\arguments{
\item{oolAttacker}{character() order of loss for attacker}

\item{oolDefender}{character() order of loss for defender}

\item{retreat}{integer() round number after which attacker should retreat, if NULL, round number will not trigger retreat}

\item{verbose}{logical() whether to print battle information to stdout}

\item{suppressChecks}{suppressChecks on input}
}
\value{
list() with members
\describe{
 \item{unitsAttacker}{remaining units for attacker, formatted as oolAttacker}
 \item{unitsDefender}{remaining units for defender, formatted as oolDefender}
 }
}
\description{
Plays one battle following the rules for Axis and Allies 1942 second edition
}
\details{
units for oolAttacker and oolDefender accepts all units specified in \code{\link[aaSimulator]{ool}}

 In addition to the standard options, the attacker and defender OOL supports the directives:
 \describe{
  \item{SUBM}{Submarines submerge at first opportunity when all units preceeding this directive is lost.}
  }
 In addition to the standard options, the attacker OOL supports the directives:
 \describe{
  \item{RET}{Attacker will retreat when all units preceeding this directive is lost.}
 }
}
\examples{
 play_1942se_battle(c("inf", "art", "ftr"), c("inf", "inf", "inf", "AA"))
}
