% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhtr2.R
\name{play_LHTR_battle}
\alias{play_LHTR_battle}
\title{Play one battle}
\usage{
play_LHTR_battle(
  oolAttacker,
  oolDefender,
  retreat = NULL,
  verbose = F,
  suppressChecks = F
)
}
\arguments{
\item{oolAttacker}{character() order of loss for attacker}

\item{oolDefender}{character() order of loss for defender}

\item{retreat}{integer() round number after which attacker should retreat, if NULL, round number will not trigger retreat}

\item{verbose}{logical() whether to print battle information to stdout}

\item{suppressChecks}{suppressChecks on input}
}
\value{
list() with members
\describe{
 \item{unitsAttacker}{remaining units for attacker, formatted as oolAttacker}
 \item{unitsDefender}{remaining units for defender, formatted as oolDefender}
 }
}
\description{
Plays one battle following Larry Harris Tournament Ruls for Axis and Allies Revised edition (LHTR 2.0)
}
\details{
units for oolAttacker and oolDefender accepts all units specified in \code{\link[aaSimulator]{ool}}

 In addition to the standard units, special units achieved through weapons development are supported:
 \describe{
  \item{Dbomb}{Destroyer offshore bombardment}
  \item{jftr}{Jet Fighter}
  \item{ssub}{Super Submarine}
  \item{hbomb}{Heavy Bomber}
 }
 Weapons development that does not result in changes to battle stats (such as long-range aircraft) does not have separate unit codes.

 In addition to the standard options, the attacker and defender OOL supports the directives:
 \describe{
  \item{SUBM}{Submarines submerge at first opportunity when all units preceeding this directive is lost.}
  }
 In addition to the standard options, the attacker OOL supports the directives:
 \describe{
  \item{RET}{Attacker will retreat when all units preceeding this directive is lost.}
 }

 LHTR 2.0 does only allow one antiaircraft gun to fire for each territory.
 Only one shot will be fired for each plane, regardless of how many aa are listed in 'oolDefender'.
 To simulate units flying over territories with AA-guns en-route to battle, this must be modelled as a separate battle.
 Battle termination conditions are check _after_ each round of battle, so it is possible ot set up a simulation of only aircrafts vs AA.
}
\examples{
 play_LHTR_battle(c("inf", "art", "ftr"), c("inf", "inf", "inf", "AA"))
}
