% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot2WaveSummary}
\alias{plot2WaveSummary}
\title{Plot summary of two-wave battle}
\usage{
plot2WaveSummary(twoWaveSimulationResults)
}
\arguments{
\item{twoWaveSimulationResults, }{formatted as: \code{\link[aaSimulator]{twoWaveSimulationResults}}}
}
\description{
Produces a panneled plot summarizing results of a two-wave battle simulation
}
\examples{
 sim <- simulateTwoWaveBattles("7 inf", "7 inf 3 ftr", "7 inf",
                               reinforcement = function(x){c(x, rep("ftr", 2))},
                               replications = 5, iterations = 1000)
 plot2WaveSummary(sim)
}
