% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{prefixedOol}
\alias{prefixedOol}
\title{string format for order of loss specification}
\description{
unit codes as for \code{\link[aaSimulator]{ool}}, but organised as a whitespace separated string rather than a vector.
}
\details{
repeted units may be prefixing units with numbers (and a whitespace to separate unit from number)
}
