% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{roll}
\alias{roll}
\title{Roll dice}
\usage{
roll(n, hitvalue)
}
\arguments{
\item{n}{number of dice to roll}

\item{hitvalue}{largest value to count as a hit}
}
\value{
numeric() number of hits
}
\description{
Rolls a set of six-sided dice, and determine number of hits.
}
