% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulateBattles}
\alias{simulateBattles}
\title{Simulate Battles}
\usage{
simulateBattles(
  oolAttacker,
  oolDefender,
  FUN = play_LHTR_battle,
  ...,
  iterations = 2000,
  replications = 3
)
}
\arguments{
\item{oolAttacker}{character() vector of unit codes in preferred order of loss}

\item{oolDefender}{character() vector of unit codes in preferred order of loss}

\item{FUN}{function for running one battle}

\item{...}{additional arguments passed to FUN}

\item{iterations}{number of iterations to simulate for each replication}

\item{replications}{number of replications to run}
}
\value{
list with one member for each replication. Each replicate contains the members:
\describe{
 \item{IPClossAttacker}{numeric() with IPC loss for attacker for each simulated battle}
 \item{IPClossDefender}{numeric() with IPC loss for defender for each simulated battle}
 \item{unitsLeftAttacker}{data.table() with units left after each simulated battle for attacker}
 \item{unitsLeftDefender}{data.table() with units left after each simulated battle for defender}
}
}
\description{
Simulate Axis and Allies Battles
}
\details{
oolAttacker and oolDefender list all units in order of preferred loss.
 In addition to regular units, some virtual units are added for signaling special abilities of units or battle choices.
 The exact list of units depends on the function provided (FUN), but the following codes are reserved, and their use is encouraged when applicable:

\describe{
  \item{inf}{Infantry}
  \item{art}{Artillery}
  \item{arm}{Tank}
  \item{ftr}{Fighter}
  \item{bmb}{Bomber}
  \item{trn}{Transport}
  \item{sub}{Submarine}
  \item{dd}{Destroyer}
  \item{ac}{Aircraft Carrier}
  \item{bb}{Battleship}
  \item{aa}{Anti-aircraft Gun. Most rule sets will allow only one per battle.}
}
 Some common virtual units:
\describe{
  \item{BBomb}{Battleship used for offshore bombardment}
  \item{BBx}{The first hit of a battleship for rulesets with two-hit capabilities. Modelled as a virtual unit with 0 in attack and defense.}
  \item{SUBM}{Flags submarines should submerge when possible.}
  \item{RET}{Flags that attacker should retreat at first opportunity when all units preceeding this virtual unit are lost.}
}
}
