% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulationStats}
\alias{simulationStats}
\title{Battle statistics for simulation}
\description{
list with the following members:
\describe{
 \item{attackerStart}{units for attacker at start of battle, formatted as oolAttacker}
 \item{defenderStart}{units for defender at start of battle, formatted as oolDefender}
 \item{avereages}{stats averaged over replicates}
 \item{replicates}{vectors with stats for all replicates of simulation}
}
}
\details{
Both 'averages' and the 'replicates' contain the stats:
\describe{
 \item{attackerWon}{proportion of battles won by attacker (all defending units lost), averaged over replicates}
 \item{defendeWon}{proportion of battles won by defender (not all defending units lost), averaged over replicates}
 \item{draw}{proportion of battles drawed (all units lost), averaged over replicates}
 \item{meanRounds}{mean number of rounds before battle was concluded}
 \item{meanAttackerCost}{mean cost of battle for attacker, averaged over iterations and replicates}
 \item{meanDefenderCost}{mean cost of battle for defender, averaged over iterations and replicates}
}
}
