% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{twoWaveSimulationResults}
\alias{twoWaveSimulationResults}
\title{Result from two-wave battle simulation}
\description{
For a simuation of n iterations, replicated m times this is:
list with members:
\describe{
 \item{firstWaveStart}{units for first wave attacker at start of battle, formatted as formatted as \code{\link[aaSimulator]{ool}}}
 \item{secondWaveStart}{units for second wave attacker at start of battle, formatted as formatted as \code{\link[aaSimulator]{ool}}}
 \item{defenderStart}{units for defender at start of battle, formatted as formatted as \code{\link[aaSimulator]{ool}}}
 \item{defenderReinforced}{units for defender at start of battle, with reinforcements added, formatted as \code{\link[aaSimulator]{ool}}}
 \item{replicates}{list of length m, with member for each replicate, each a list of length n containing two members: 'firstWave' and 'secondWave', both formatted as \code{\link[aaSimulator]{battleResults}}}
}
}
