% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1942se_units_doc.R
\docType{data}
\name{unitlist1942se}
\alias{unitlist1942se}
\title{Table of units allowed in Axis and Allies 1942 Second Edition}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 18 rows and 8 columns.
}
\usage{
data(unitlist1942se)
}
\description{
Formatted as \code{\link[aaSimulator]{unitTable}}
}
\keyword{datasets}
