% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{makePosteriorDistributionPlot}
\alias{makePosteriorDistributionPlot}
\title{plot posterior distribution}
\usage{
makePosteriorDistributionPlot(
  simulationResults,
  side = "attacker",
  attackCol = "red",
  defenceCol = "blue",
  ylim = c(0, 1)
)
}
\arguments{
\item{simulationResults, }{formatted as: \code{\link[aaSimulator]{simulationResults}}}

\item{side}{character(), side to calculate distribution for, either 'attacker' or 'defender'}

\item{attackCol}{color to use when side is 'attacker'}

\item{defenceCol}{color to use when side is 'defender'}

\item{ylim}{limits for y-axis}
}
\value{
ggplot object
}
\description{
Calculate and plot posterior distribution for battle
}
\details{
Plotter error bars are the ranges of the reported probabilities across replicates.
}
\examples{
 sim <- simulateBattles("30 inf 11 arm 5 ftr arm", "40 inf 3 ftr", iterations = 2000, replications = 3)
 makePosteriorDistributionPlot(sim, "attacker")
 makePosteriorDistributionPlot(sim, "defender")
}
