% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhtr2.R
\name{play_LHTR_battle}
\alias{play_LHTR_battle}
\title{Play one battle}
\usage{
play_LHTR_battle(
  oolAttacker,
  oolDefender,
  retreat = NULL,
  verbose = F,
  suppressChecks = F
)
}
\arguments{
\item{oolAttacker}{character() order of loss for attacker}

\item{oolDefender}{character() order of loss for defender}

\item{retreat}{integer() round number after which attacker should retreat, if NULL, round number will not trigger retreat}

\item{verbose}{logical() whether to print battle information to stdout}

\item{suppressChecks}{suppressChecks on input}
}
\value{
list() with members
\describe{
 \item{unitsAttacker}{remaining units for attacker, formatted as oolAttacker}
 \item{unitsDefender}{remaining units for defender, formatted as oolDefender}
 }
}
\description{
Plays one battle following Larry Harris Tournament Ruls for Axis and Allies Revised edition (LHTR 2.0)
}
\details{
In addition to the standard options, the attacker and defender OOL supports the directives:
 \describe{
  \item{+sumberge}{Submarines submerge at first opportunity (after any round of combat with no enemy destroyer present)}
  }
 In addition to the standard options, the attacker OOL supports the directives:
 \describe{
  \item{+roundretreat <n>}{Attacker will retreat after round n.}
  \item{+ret}{Attacker will retreat when all units preceeding this directive is lost.}
 }
}
