% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotBattleSummary}
\alias{plotBattleSummary}
\title{Plot battle summary}
\usage{
plotBattleSummary(simulationResults)
}
\arguments{
\item{simulationResults, }{formatted as: \code{\link[aaSimulator]{simulationResults}}}
}
\description{
Produces a panneled plot summarizing results of a battle simulation
}
\examples{
 sim <- simulateBattles("30 inf 11 arm 5 ftr arm",
        "40 inf 3 ftr",
        iterations = 1000,
        replications = 2)
 plotBattleSummary(sim)
}
