% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulateBattles}
\alias{simulateBattles}
\title{Simulate Battles}
\usage{
simulateBattles(
  oolAttacker,
  oolDefender,
  FUN = play_LHTR_battle,
  ...,
  iterations = 2000,
  replications = 3
)
}
\arguments{
\item{oolAttacker}{character() vector of unit codes in preferred order of loss, formatted as \code{\link[aaSimulator]{ool}}}

\item{oolDefender}{character() vector of unit codes in preferred order of loss, formatted as \code{\link[aaSimulator]{ool}}}

\item{FUN}{function for running one battle}

\item{...}{additional arguments passed to FUN}

\item{iterations}{number of iterations to simulate for each replication}

\item{replications}{number of replicates to run}
}
\value{
\code{\link[aaSimulator]{simulationResults}}
}
\description{
Simulate Axis and Allies Battles
}
\details{
The battle function 'FUN' implements the ruleset used.
FUN must accept the arguments 'oolAttacker' and 'oolDefender' and a logical argument suppressChecks,
it must return \code{\link[aaSimulator]{battleResults}}
}
