% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulationResults}
\alias{simulationResults}
\title{Result from battle simulation}
\description{
For a simuation of n iterations, replicated m times this is:
list with members:
\describe{
 \item{attackerStart}{units for attacker at start of battle, formatted as oolAttacker}
 \item{defenderStart}{units for defender at start of battle, formatted as oolDefender}
 \item{replicates}{list of length m, with member for each replicate, each a list of length n containing \code{\link[aaSimulator]{battleResults}}}
}
}
